Added tests.
Switched from run_test to tox, as I wanted to do a few non unit test
things with it, and that seemed simplest.

Unit tests are very basic. Run with `tox`.

Also has a deploy script that can be run with `tox -e deploy`, which
will destructively deploy microk8s to a model on the current
controller.