Code reorg and README update.
Moved shell stuff into a different file, renaming things and updating
tests to match.

The intent is to make it more straightforward to tell what the charm
is doing by reading charm.py, and get rid of some of the fiddly param
passing of calling a 'shell' command. Instead, we can call 'check' and
'check_output', which make more semantic sense.